pdf-to-sentences/requirements.txt
fastapi
uvicorn[standard]
requests
markitdown[pdf]
python-multipart
marker-pdf
spacy